[ARM; ts7400] Remove regulators for RS232, MODBUS, etc.
Regulators sound like the right tool for the job but:
- they can't be controlled from userspace
- the serial driver's don't know how to turn them on and off

The RS-232 ports have a transceiver that is switched by GPIO 1/25.
The Modbus port has some pins that output 24V when GPIOs 1/13 and 1/15
are both turned on.  There's also a status pin to indicate a fault on
GPIO 1/14.

We'll just leave the GPIOs unregistered so they can be accessed via
userspace.