ARM: ts7400: Disable spi-gpio for companion IC comms
It seems someone goofed up with the design of the modem daughter card
and wired up some of the modems (notably reset and RTS) to the SPI
signals going to the Cortex M0 companion chip.

Consequently, I've had to disable this feature and will have to look at
what signals being asserted on the GPIO lines cause the modem to break.